refactor(capture): extract ffmpeg/encoder logic into ffmpeg.js
Moves the encoder profiles, output-arg construction, and a spawn helper out of
screencast.js into a focused leaf module (src/ffmpeg.js), mirroring src/ebml.js.
screencast.js now reads as: spawn ffmpeg, mux CDP frames via ebml, collect the
output. No behaviour change; tests moved to test/ffmpeg.js.
