+++ packages/capture/src/index.js
@@ -12,6 +12,7 @@ const {
 } = require('./constants')
 const runCapture = require('./capture')
 const runScreencastCapture = require('./screencast')
+const { ENCODERS } = require('./ffmpeg')
 
 const BACKENDS = { extension: runCapture, screencast: runScreencastCapture }
 const DEFAULT_BACKEND = 'extension'
@@ -53,6 +54,6 @@ module.exports.extensionPath = EXTENSION_PATH
 module.exports.extensionId = EXTENSION_ID
 module.exports.TYPES = TYPES
 module.exports.BACKENDS = Object.keys(BACKENDS)
-module.exports.ENCODERS = runScreencastCapture.ENCODERS
+module.exports.ENCODERS = ENCODERS
 module.exports.DEFAULT = DEFAULT
 module.exports.DEFAULT_CODEC_BY_TYPE = DEFAULT_CODEC_BY_TYPE
